chore(mailchimp): move everything related to mailchimp from mail to mailchimp module
