+++ apps/api/package.json
@@ -27,6 +27,7 @@
     "@orbiting/backend-modules-keyCDN": "*",
     "@orbiting/backend-modules-lead-tracking": "*",
     "@orbiting/backend-modules-mail": "*",
+    "@orbiting/backend-modules-mailchimp": "*",
     "@orbiting/backend-modules-mailbox": "*",
     "@orbiting/backend-modules-maillog": "*",
     "@orbiting/backend-modules-matomo": "*",
