+++ apps/assets/package.json
@@ -26,6 +26,7 @@
     "@orbiting/backend-modules-invoices": "*",
     "@orbiting/backend-modules-keyCDN": "*",
     "@orbiting/backend-modules-mail": "*",
+    "@orbiting/backend-modules-mailchimp": "*",
     "@orbiting/backend-modules-mailbox": "*",
     "@orbiting/backend-modules-matomo": "*",
     "@orbiting/backend-modules-migrations": "*",
