+++ packages/backend-modules/mail/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@orbiting/backend-modules-mail",
   "version": "2.0.0",
-  "description": "simple mails with mandrill, manage subscriptions on mailchimp",
+  "description": "simple mails with mandrill",
   "main": "index.js",
   "repository": {
     "type": "git",
