+++ packages/backend-modules/mailchimp/NewsletterSubscription.u.jest.js
