+++ packages/backend-modules/mailchimp/index.ts
@@ -1,1 +1,3 @@
-
+module.exports = {
+  ...require('./lib'),
+}
