+++ packages/backend-modules/mailchimp/lib/deleteEmail.js
@@ -1,5 +1,5 @@
 const MailchimpInterface = require('../MailchimpInterface')
-const { EmailRequiredMailError } = require('../errors')
+const { EmailRequiredMailError } = require('@orbiting/backend-modules-mail/errors')
 const logger = console
 
 module.exports = async ({ email }) => {
