+++ packages/backend-modules/mailchimp/lib/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  changeEmailOnMailchimp: require('./changeEmailOnMailchimp'),
+  deleteEmail: require('./deleteEmail'),
+  getNewsletterSettings: require('./getNewsletterSettings'),
+  updateMergeFields: require('./updateMergeFields'),
+  updateNewsletterSubscriptions: require('./updateNewsletterSubscriptions'),
+}
