+++ packages/backend-modules/mailchimp/script/mailchimp/runOperations.js
